pyproject.toml
[tool.poetry]
name = "test-seazone-code"
version = "0.1.0"
description = ""
authors = ["Léo Graciano <leonardoferreiragraciano@gmail.com>"]
readme = "README.md"
packages = [{include = "test_seazone_code"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2"
django-filter = "^22.1"
djangorestframework = "^3.14.0"
python-decouple = "^3.6"
faker = "^16.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
